Create AWSBucketConn to hold aws functionality
