+++ ext/storage_sources/s3_store/AwsBucketConn.cpp
@@ -1,27 +1,31 @@
 #include <aws/core/Aws.h>
 #include <aws/s3-crt/S3CrtClient.h>
 #include <iostream>
+#include "AwsBucketConn.h"
+ 
+AwsBucketConn::AwsBucketConn(Aws::S3Crt::ClientConfiguration config, Aws::SDKOptions options)
+    : m_aws_config(config), m_options(options){}
 
+void AwsBucketConn::initBucketConn() {
+    Aws::InitAPI(m_options);
+    Aws::S3Crt::S3CrtClient *s3_crt_client;
+    s3_crt_client = new Aws::S3Crt::S3CrtClient(m_aws_config);
+    m_s3_crt_client = *s3_crt_client;
+}
 
-class AwsBucketConn{
+bool AwsBucketConn::s3_list_buckets()
+{
+    Aws::S3Crt::Model::ListBucketsOutcome outcome = m_s3_crt_client.ListBuckets();
 
-    private:
-    Aws::S3Crt::ClientConfiguration aws_config;
-    Aws::S3Crt::S3CrtClient s3_crt_client;
-    Aws::String region = Aws::Region::US_EAST_1;
-    const double throughput_target_gbps = 5;
-    const uint64_t part_size = 8 * 1024 * 1024; // 8 MB.
-    
-
-
-
-    public: 
-    void ListBuckets(){
-
+    if (outcome.IsSuccess()) {
+        std::cout << "All buckets under my account:" << std::endl;
+        for (auto const &bucket : outcome.GetResult().GetBuckets()) {
+            std::cout << "  * " << bucket.GetName() << std::endl;
+        }
+        std::cout << std::endl;
+        return true;
+    } else {
+        std::cout << "ListBuckets error:\n" << outcome.GetError() << std::endl << std::endl;
+        return false;
     }
-};
-
-
-int main(){
-    AwsBucketConn conn1;
 }
